+++ shoezam/scraper/enums.py
@@ -1,0 +1,13 @@
+from enum import Enum
+
+class Gender(Enum):
+    MEN = 'men'
+    WOMEN = 'women'
+    KIDS = 'kids'
+
+
+class ShoeType(Enum):
+    OXFORDS = 'oxfords'
+    SNEAKER = 'sneakers-athletic-shoes'
+
+
